requirements.txt
pandas
seaborn
plotly
requests
scikit-learn
mysql-connector-python
xgboost
matplotlib
sort-dataframeby-monthorweek
sorted_months_weekdays
jinja2
fastapi
uvicorn[standard]
databases
sqlalchemy
pyodbc


beautifulsoup4